# Make global variables constant