--- a/src/Convex.jl
+++ b/src/Convex.jl
@@ -84,9 +84,10 @@ include("utilities/broadcast.jl")
 
 #Temporary workaround for memory leak (https://github.com/JuliaOpt/Convex.jl/issues/83)
 function clearmemory()
-    global id_to_variables = Dict{UInt64, Variable}()
-    global var_to_ranges = Dict{UInt64, Tuple{Int, Int}}()
-    global conic_constr_to_constr = Dict{ConicConstr, Constraint}()
+    global id_to_variables
+    empty!(id_to_variables)
+    global conic_constr_to_constr 
+    empty!(conic_constr_to_constr)
     GC.gc()
 end