@ericphanson 2019-05-08
Is there a good way to add a test for `clearmemory`?