@ararslan 2019-05-08
> Is there a good way to add a test for `clearmemory`?

I guess one way would be to populate those global variables with something then call `clearmemory` and test that they're empty.