@ericphanson 2019-05-08
Ok, I tried adding a test. I just realized though maybe you meant I should populate them by hand. Let me know if this is fine or if I should change it.